% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{adjust_shot_end_coords}
\alias{adjust_shot_end_coords}
\title{Adjust shot end coordinates}
\usage{
adjust_shot_end_coords(data)
}
\arguments{
\item{data}{A data frame with shot details and the original shot end coordinates}
}
\value{
The same data frame with corrected shot end coordinate columns
}
\description{
Adjust the shot end coordinates to fix bias in data collection process for shots
near the posts.
}
