% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{clean_raw_statsbomb_data}
\alias{clean_raw_statsbomb_data}
\title{Clean raw StatsBomb data}
\usage{
clean_raw_statsbomb_data(data)
}
\arguments{
\item{data}{A data frame with shot details, and the original columns from StatsBomb data}
}
\value{
The data frame modified for use in the shooting skill modeling pipeline
}
\description{
Reformat columns from raw StatsBomb shot data
}
