% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution_error_linear_fits.R
\docType{data}
\name{execution_error_linear_fits}
\alias{execution_error_linear_fits}
\title{Linear fits for execution error}
\format{
An object of class \code{rowwise_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3 rows and 3 columns.
}
\usage{
execution_error_linear_fits
}
\description{
A data frame with the model column containing lm models for each of the three
components of the covariance matrix for execution error.
Based on data \link{Hunter_et_al_2018_shots}.
}
\keyword{datasets}
