% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_shot_xg.R
\name{filter_post_xg_shots}
\alias{filter_post_xg_shots}
\title{Filter shots for post-shot expected goals model}
\usage{
filter_post_xg_shots(data)
}
\arguments{
\item{data}{A data frame with shot details}
}
\value{
The same data with only shots used to train the post-shot expected goals model.
}
\description{
Filter the shots data to include only shots used to train the
post-shot expected goals model.
}
