% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{filter_shooting_skill_data}
\alias{filter_shooting_skill_data}
\title{Filter shooting skill data}
\usage{
filter_shooting_skill_data(data)
}
\arguments{
\item{data}{A data frame with shot details}
}
\value{
The same data frame with some rows removed
}
\description{
Remove rows from the shots data that are not used to measure shooting skill.
Specifically, keep only shots that are at least 15 yards from the goal and have a StatsBomb
pre-shot expected goals value of less than 0.1.
}
