% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_model.R
\name{fit_global_weights}
\alias{fit_global_weights}
\title{Fit global component weights}
\usage{
fit_global_weights(pdfs)
}
\arguments{
\item{pdfs}{An n x m matrix with the (i, j) entry containing the probability
density function of the i'th shot for the j'th mixture model component.}
}
\value{
A vector of length m indicating the mixture weights of each component.
}
\description{
Given a matrix containing the probability density functions for each shot and
component pair, fit the mixture model weights of the component for the global
mixture model.
}
