% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_model.R
\name{fit_player_weights}
\alias{fit_player_weights}
\title{Fit player component weights}
\usage{
fit_player_weights(pdfs, player_labels, global_weights, alpha = 30)
}
\arguments{
\item{pdfs}{An n x k matrix with the (i, j) entry containing the probability
density function of the i'th shot for the j'th mixture model component.}

\item{player_labels}{An integer vector of length n indicating the player associated with
each shot. If there are p players, player_labels should contain only the
integers 1 through p.}

\item{global_weights}{A vector of length k indicating the weights of each
component for the global mixture model.}

\item{alpha}{A number representing the degree to which player weights are
shrunk towards the global weights. alpha = 0 corresponds to no shrinkage,
while alpha = Inf will force all players to have the global weights.}
}
\value{
A p x k matrix with the (i, j) entry corresponding to the weight of
the j'th component for player i.
}
\description{
Given a matrix containing the probability density functions for each shot and
component pair, fit the mixture model weights of the component for the global
mixture model.
}
