% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution_error.R
\name{flip_left_foot_Hunter_shots}
\alias{flip_left_foot_Hunter_shots}
\title{Flip left-footed shots from Hunter et al. 2018}
\usage{
flip_left_foot_Hunter_shots(data)
}
\arguments{
\item{data}{A data frame with each row representing one shot, with "foot" and
"horizontal_error_yards" columns}
}
\value{
The same data frame with the "horizontal_error_yards" column modified
}
\description{
Flip the horizontal error for left-footed shots
}
