% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{flip_left_foot_shot_end_coords}
\alias{flip_left_foot_shot_end_coords}
\title{Flip left-footed shot end coordinates}
\usage{
flip_left_foot_shot_end_coords(data)
}
\arguments{
\item{data}{A data frame with shot details and the original shot end coordinates}
}
\value{
The same data frame with the \code{y_end_proj} column changed.
}
\description{
Reflect the shot end coordinates for left-footed shots about the centre of the net.
}
