% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_model.R
\name{get_component_values}
\alias{get_component_values}
\title{Get component values}
\usage{
get_component_values(mixture_model_components, n_samples = 100)
}
\arguments{
\item{mixture_model_components}{A data.frame with columns mean and cov representing
the mean vector and covariance matrix for each component.}

\item{n_samples}{The number of samples from each component used for Monte Carlo integration}
}
\value{
The same data frame with new column "value" indicating the mean post-shot
expected goals value for a shot samples from that component.
}
\description{
Get the expected value of each truncated bivariate Gaussian distribution,
computed using Monte Carlo simulation.
}
