% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution_error.R
\name{get_execution_error_covariance}
\alias{get_execution_error_covariance}
\title{Get covariance matrix of execution error}
\usage{
get_execution_error_covariance(height)
}
\arguments{
\item{height}{A numeric value, the intended shot height in yards}
}
\value{
A 2x2 covariance matrix
}
\description{
Get the covariance matrix of exeucution error in the (y, z) end coordinates
for soccer shots aimed at a given height.
}
\examples{
get_execution_error_covariance(1.25)

}
