% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_model.R
\name{get_mixture_model_components}
\alias{get_mixture_model_components}
\title{Get bivariate Gaussian mixture model components}
\usage{
get_mixture_model_components(num_y = 11, num_z = 6, lambdas = c(1, 3.8))
}
\arguments{
\item{num_y}{The number of grid points in the y direction}

\item{num_z}{The number of grid points in the z direction}

\item{lambdas}{A numeric vector containing the scaling factors for each layer}
}
\value{
Data frame with columns mean, lambda, and cov indicating the component
mean, associated lambda value for scaling, and (scaled) covariance matrix.
}
\description{
The means of the components are evenly-spaced points along a grid within the goal frame.
The covariance matrices of the components are based on empirical observations of
the shape of execution errors for different shot target heights.
There are multiple layers of components, with each layer having a different scaling factor lambda.
}
