% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{get_player_groups}
\alias{get_player_groups}
\title{Get player groups}
\usage{
get_player_groups(
  shots_data,
  grouping_cols = c("player"),
  group_size_threshold = 30
)
}
\arguments{
\item{shots_data}{A dataframe with one row per shot}

\item{grouping_cols}{A character vector of columns in shots_data, indicating
the level(s) at which to fit mixture model components. Typically, this will be
at the player, player-season, or player-half-season level.}

\item{group_size_threshold}{An integer indicating the minimum group size to be included.}
}
\value{
shots_data with two modifications. First, a new column group_id is
added, containing integers from 1 to some number k identifying the group for each shot.
Second, the dataframe is filtered to only groups meeting the specified threshold.
}
\description{
Add a column indicating the index of the group each shot belongs to,
and filter to groups meeting a specified threshold.
}
