% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_model.R
\name{get_shot_probability_densities}
\alias{get_shot_probability_densities}
\title{Get shot probability densities}
\usage{
get_shot_probability_densities(mixture_model_components, shots)
}
\arguments{
\item{mixture_model_components}{A data.frame with m rows and columns mean and
cov representing the mean vector and covariance matrix for each component.}

\item{shots}{An n x 2 matrix with the columns representing the y- and z-coordinates respecitvely.}
}
\value{
An n x m matrix with the (i, j) entry containing the probability density
function of the i'th shot for the j'th mixture model component.
}
\description{
Get a matrix with the density function for each shot and each component.
}
