% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_weights.R
\docType{data}
\name{global_weights}
\alias{global_weights}
\title{Global mixture model weights fit via Stan}
\format{
An object of class \code{numeric} of length 8.
}
\usage{
global_weights
}
\description{
The global mixture model weights from the second stage as fit by
variational inference via Stan.
}
\keyword{datasets}
