% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hunter_et_al_2018_shots.R
\docType{data}
\name{Hunter_et_al_2018_shots}
\alias{Hunter_et_al_2018_shots}
\title{Data from Hunter et al. 2018}
\format{
An object of class \code{data.frame} with 8466 rows and 4 columns.
}
\source{
Kindly provided by Dr. Andrew Hunter.
}
\usage{
Hunter_et_al_2018_shots
}
\description{
The full shot data used in the paper. Available upon request from a.hunter@uq.edu.au.
}
\keyword{datasets}
