% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_on_target}
\alias{is_on_target}
\title{Is shot on target?}
\usage{
is_on_target(y, z)
}
\arguments{
\item{y}{A numeric vector of shot end y coordinates}

\item{z}{A numeric vector of shot end z coordinates}
}
\value{
Boolean vector indicating whether each shot is on target
}
\description{
Is shot on target?
}
