% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{load_gen_post_xg}
\alias{load_gen_post_xg}
\title{Load GenPostXg}
\usage{
load_gen_post_xg(shooting_skill_data, pdfs, mixture_model_components)
}
\arguments{
\item{shooting_skill_data}{A data frame with shot details and a \code{group_id}
column indicating the grouping level at which player weights were fit.}

\item{pdfs}{Output from \link{get_shot_probability_densities}.}

\item{mixture_model_components}{Output from \link{get_mixture_model_components}.}
}
\value{
The same shots data frame with a new column \code{gen_post_xg}.
}
\description{
Adds a column \code{gen_post_xg} to a data frame of shots given the component weights
and mixture model components.
}
