% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{load_rb_post_xg}
\alias{load_rb_post_xg}
\title{Load RBPostXg}
\usage{
load_rb_post_xg(shooting_skill_data, player_weights, component_values)
}
\arguments{
\item{shooting_skill_data}{A data frame with shot details and a \code{group_id}
column indicating the grouping level at which player weights were fit.}

\item{player_weights}{A p x k matrix with each row corresponding to the mixture
model component weights associated with group p. Output from \link{fit_player_weights}.}

\item{component_values}{Vector of length k indicating the value of each
mixture model component.}
}
\value{
The same shots data frame with a new column \code{rb_post_xg}. The \code{rb_post_xg}
values will be identical for any rows with identical \code{group_id}.
}
\description{
Adds a column \code{rb_post_xg} to a data frame of shots given the component weights
and mixture model components.
}
