% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_weight_posteriors.R
\docType{data}
\name{player_weight_posteriors}
\alias{player_weight_posteriors}
\title{Posteriors from fitting player weights via MCMC}
\format{
An object of class \code{list} of length 3.
}
\usage{
player_weight_posteriors
}
\description{
A list generated via rstan::extract(stanfit) including all the posterior
draws for the global weights and player weights.
}
\keyword{datasets}
