% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_weights.R
\docType{data}
\name{player_weights}
\alias{player_weights}
\title{Player mixture model weights fit via Stan}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2914 rows and 8 columns.
}
\usage{
player_weights
}
\description{
The mixture model weights at the player-half-season level as fit by
variational inference via Stan.
}
\keyword{datasets}
