% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_goalposts}
\alias{plot_goalposts}
\title{Plot goal posts}
\usage{
plot_goalposts(...)
}
\arguments{
\item{...}{Arguments passed to geom_segment}
}
\value{
ggplot2::geom_segment with both posts and crossbar
}
\description{
Add line segments showing the goal posts and crossbar to a ggplot.
}
