% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_grass}
\alias{plot_grass}
\title{Plot grass}
\usage{
plot_grass(...)
}
\arguments{
\item{...}{Arguments passed to geom_hline}
}
\value{
ggplot2::geom_hline at y=0
}
\description{
Add horizontal line segment at y=0 to ggplot.
}
