% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_shot_xg.R
\docType{data}
\name{post_shot_xg_model}
\alias{post_shot_xg_model}
\title{Coordinates-based Post-shot Expected Goals Model}
\format{
An object of class \code{glm} (inherits from \code{lm}) of length 21.
}
\usage{
post_shot_xg_model
}
\description{
The post-shot expected goals model we develop based off only the shot end coordinates
}
\keyword{datasets}
