% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_shot_xg.R
\name{predict_post_xg}
\alias{predict_post_xg}
\title{Predict post-shot expected goals}
\usage{
predict_post_xg(y, z)
}
\arguments{
\item{data}{A data frame with shot details}

\item{model}{A fit glm model used to predict post-shot expected goals values}
}
\value{
The same data with only shots used to train the post-shot expected goals model.
}
\description{
Predict the probability of a goal given the shot details and
post-shot expected goals model.
}
\examples{
predict_post_xg(y = 35:42, z = 0:3)
}
