% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{prepare_shooting_skill_data}
\alias{prepare_shooting_skill_data}
\title{Prepare shooting skill data}
\usage{
prepare_shooting_skill_data(data)
}
\arguments{
\item{data}{A data frame with shot details}
}
\value{
The data frame modified for use in the shooting skill modeling pipeline
}
\description{
Prepare data used to measure shooting skill. Some rows are removed
and left-footed shots are reflected.
}
