% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{project_shot_end_coords}
\alias{project_shot_end_coords}
\title{Project shot end coordinates}
\usage{
project_shot_end_coords(data)
}
\arguments{
\item{data}{A data frame with shot details and the original shot end coordinates}
}
\value{
The same data frame with new columns \code{y_end_proj} and \code{z_end_proj}
}
\description{
Project the shot end coordinates for shots with \code{x_end} != 120,
primarily saved shots. Uses projectile motion equations to impute the shot
trajectory and infer the shot end location.
}
