% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_components.R
\docType{data}
\name{selected_components}
\alias{selected_components}
\title{Indices of components selected via pruning}
\format{
An object of class \code{integer} of length 8.
}
\usage{
selected_components
}
\description{
The indices of the full mixture model components kept after pruning components
based on their mixture weights.
}
\keyword{datasets}
