% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shot_metrics.R
\docType{data}
\name{shot_metrics}
\alias{shot_metrics}
\title{Shot-level metrics}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21010 rows and 24 columns.
}
\usage{
shot_metrics
}
\description{
A \code{data.frame} containing shot-specific information indlucing our
proposed metrics.
}
\keyword{datasets}
