% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statsbomb_shots}
\alias{statsbomb_shots}
\title{StatsBomb shots data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 118541 rows and 18 columns.
}
\source{
Provided by StatsBomb for our academic research collaboration with Toronto FC.
}
\usage{
statsbomb_shots
}
\description{
The full shot data used in the paper. Not available to the public.
}
\keyword{datasets}
