% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{filter_shooting_skill_data}
\alias{filter_shooting_skill_data}
\title{Filter shooting skill data}
\usage{
filter_shooting_skill_data(data, max_xg = 0.1, min_distance = 15)
}
\arguments{
\item{data}{A data frame with shot details}

\item{max_xg}{A number indicating the maximum pre-shot expected goals to include}

\item{min_distance}{A number indicating the minimum shot distance to include}
}
\value{
The same data frame with some rows removed
}
\description{
Remove rows from the shots data that are not used to measure shooting skill.
Specifically, keep only shots from at least 15 yards out that have a StatsBomb
pre-shot expected goals value of less than 0.1 and valid shot end coordinates.
}
