% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hunter_et_al_2018_shots.R
\docType{data}
\name{get_execution_error_covariance}
\alias{get_execution_error_covariance}
\title{Get covariance matrix of execution error}
\format{
An object of class \code{function} of length 1.
}
\usage{
get_execution_error_covariance(height)

get_execution_error_covariance
}
\arguments{
\item{height}{The shot target height in yards}
}
\value{
A 2x2 covariance matrix

A 2x2 covariance matrix representing the estimated execution error for
a shot aimed at the given target height, with all units in yards
}
\description{
Get the covariance matrix of exeucution error in the (y, z) end coordinates
for soccer shots aimed at a given height.
}
\examples{
get_execution_error_covariance(1.25)
}
\keyword{datasets}
