% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{prepare_shooting_skill_data}
\alias{prepare_shooting_skill_data}
\title{Prepare shooting skill data}
\usage{
prepare_shooting_skill_data(data, max_xg = 0.1, min_distance = 15)
}
\arguments{
\item{data}{A data frame with shot details}

\item{max_xg}{A number indicating the maximum pre-shot expected goals to include}

\item{min_distance}{A number indicating the minimum shot distance to include}
}
\value{
The data frame modified for use in the shooting skill modeling pipeline
}
\description{
Prepare data used to measure shooting skill. Some rows are removed
and left-footed shots are reflected.
}
