% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsbomb_shots_processed.R
\docType{data}
\name{statsbomb_shots_processed}
\alias{statsbomb_shots_processed}
\title{Processed StatsBomb shots data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35610 rows and 20 columns.
}
\source{
Based on \code{statsbomb_shots}
}
\usage{
statsbomb_shots_processed
}
\description{
The processed full shot data used in the paper. Not available to the public.
}
\keyword{datasets}
